fix integer overflow in sum_of_squared_gaussians
Changed argument from int to int64_t to avoid overflow for
systems with more than 2 billion degrees of freedom.

+++ src/Random.h
@@ -175,7 +175,7 @@ public:
   *  \return a random variable equal to the sum of num_gaussians squared
   *    standard normal variables
   */
-  BigReal sum_of_squared_gaussians(int num_gaussians) {
+  BigReal sum_of_squared_gaussians(int64_t num_gaussians) {
     BigReal z, u, v;
 
     if (num_gaussians <= 2) {
